Add baseline .idea/vcs.xml file
This is being added in anticipation of changes that will further modify
this file to configure IDEA's Git commit dialog settings to conform to
the seven rules of How to Write a Git Commit Message.

In the meantime, this change should be helpful on its own, as it will
avoid the annoying popup about "Unmapped Git VCS roots" after first
import.